chore(ds): dehookify GraphQL related functions
